pyproject.toml
[tool.poetry]
name = "blocknotify-net"
version = "0.1.0"
description = "A notification service that monitors blockchain events."
authors = ["Nik <nik@pyxis.com>"]
readme = "README.md"
#packages = [{include = "blocknotify"}]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.7"
dj-database-url = "^1.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
